renamed app to app.py (I was lazy while launching the app). Added Map and Room classes, the basic gameplay loop should now work, but needs to be made prettier.
